requirements.txt
mujoco>=3.0.0
mujoco-python-viewer>=0.1.0
numpy>=1.24.0
torch>=2.0.0
PyYAML>=6.0
